requirements_airflow.txt
apache-airflow
pycaret[full]
mlflow
scikit-learn
pandas
psycopg2-binary